requirements.txt
experta
pandas==2.1.0
numpy==1.24.3
Flask==2.3.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0 